requirements.txt
jupyterlab==3.0.16
wandb==0.11.0
scikit-learn
matplotlib
overrides
contexttimer
pandas
click
graphviz
overrides
mypy
transformers==4.5.1
spacy
datasets
rouge-score
absl-py
line_profiler
h5py
ipywidgets
hydra-core
pytorch-lightning
sacrebleu
sentence-transformers
pytest